test(halo2_ivc): extend positive golden coverage and refactor asset workflow
