+++ mithril-stm/src/circuits/halo2_ivc/tests/golden/mod.rs
@@ -1,6 +1,6 @@
 //! Golden tests lock in recursive Halo2 IVC verification behavior.
 //! This subtree also owns the committed assets, asset readers, generators, and
-//! scenario/docs for the golden workflow.
+//! supporting test infrastructure for the golden workflow.
 
 pub(crate) const ASSET_SEED: u64 = 42;
 pub(crate) const CERTIFICATE_CIRCUIT_DEGREE: u32 = 13;
